Refactor ALICE download code for clarity
There are no code functionality changes in this commit (except for
adapting the `__init__.py`).
