+++ pachyderm/alice/__init__.py
@@ -6,7 +6,9 @@
 """
 
 __all__ = [
-    "utils"
+    "download",
+    "utils",
 ]
 
+from .download import download_dataset, download_run_by_run_train_output  # noqa: F401
 from .utils import copy_from_alien, grid_md5, list_alien_dir  # noqa: F401
